app/requirements.txt
fastapi==0.115.12
Pillow==11.2.1
protobuf==6.31.1
pydantic==2.11.5
PyJWT==2.10.1
python_bcrypt==0.3.2
SQLAlchemy==2.0.35
uvicorn==0.34.3
slowapi
pyparsing==3.3.1
phonenumbers==9.0.21
python-dotenv
google-generativeai
pytz
pyotp